more idiomatic: private module
